Move persistence providers code into util since they depend on ember.
Unit tests pass, will need to fix up apps.
