+++ src/app/SafeAttributePersistenceProvider.h
@@ -18,11 +18,12 @@
 #include <app/ConcreteAttributePath.h>
 #include <app/data-model/Nullable.h>
 #include <app/util/attribute-metadata.h>
-#include <cstring>
-#include <inttypes.h>
 #include <lib/support/BufferReader.h>
 #include <lib/support/BufferWriter.h>
 #include <lib/support/Span.h>
+
+#include <cstring>
+#include <inttypes.h>
 
 namespace chip {
 namespace app {
