+++ src/app/util/persistence/BUILD.gn
@@ -1,0 +1,47 @@
+# Copyright (c) 2024 Project CHIP Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import("//build_overrides/chip.gni")
+
+source_set("persistence") {
+  sources = [
+    "AttributePersistenceProvider.h",
+    "DefaultAttributePersistenceProvider.cpp",
+    "DefaultAttributePersistenceProvider.h",
+  ]
+
+  public_deps = [
+    "${chip_root}/src/app:attribute-persistence",
+    "${chip_root}/src/app:paths",
+    "${chip_root}/src/app/util:types",
+    "${chip_root}/src/lib/core",
+    "${chip_root}/src/lib/support",
+    "${chip_root}/src/lib/support:span",
+  ]
+}
+
+source_set("deferred") {
+  sources = [
+    "DeferredAttributePersistenceProvider.cpp",
+    "DeferredAttributePersistenceProvider.h",
+  ]
+
+  public_deps = [
+    ":persistence",
+    "${chip_root}/src/lib/support",
+    "${chip_root}/src/lib/support:span",
+    "${chip_root}/src/platform",
+    "${chip_root}/src/system",
+  ]
+}
